update
can delete reports, lifeTime bar complete [ Dummy data ]

+++ lib/provider/parkingLot.dart
@@ -16,7 +16,7 @@ class ParkLot with ChangeNotifier {
   @required
   List poly;
   @required
-  List<double> lifeTime;
+  List<int> lifeTime;
 
   ParkLot({
     this.id,
